extract mocks from rmc (mostly)
