+++ mock/src/lib.rs
@@ -2,12 +2,16 @@ mod crypto;
 mod dataio;
 mod hasher;
 mod network;
+mod rmc;
 mod spawner;
+mod threshold_multi_keychain;
 mod unreliable_router;
 
-pub use crypto::{KeyBox, PartialMultisignature, Signature};
+pub use crypto::{KeyBox, PartialMultisignature, Signature, VerboseKeyBox, VerboseSignature};
 pub use dataio::{Data, DataProvider, FinalizationHandler};
 pub use hasher::{Hash64, Hasher64};
 pub use network::{Network, NetworkReceiver, NetworkSender};
+pub use rmc::*;
 pub use spawner::Spawner;
+pub use threshold_multi_keychain::DefaultMultiKeychain;
 pub use unreliable_router::{configure_network, NetworkHook, Peer, UnreliableRouter};
